wip:
The problem here is that if the TS are narrowed and a new reqid is
allocated the temporary SA is not removed.  So traffic that doesn't
match the established SA won't trigger an acquire until that SA expires.
Narrowing works (e.g. in the trap-any case) if the reqid is kept constant
and selectors are configured on the SAs. Not sure why there was this
lookup previously, perhaps when a trapped SA was triggered manually?
However, if there is narrowing in that case (and a new reqid) that
shouldn't be a problem.