child-cfg: Strip DH groups from both compared proposals if requested
This fixes two issues, one is a bug if a DH group is configured for the
local ESP proposals and charon.prefer_configured_proposals is disabled.
This would cause the DH groups to get stripped not from the configured but
from the supplied proposal, which usually already has them stripped.  So
the proposals wouldn't match.  We'd have to always strip them from the local
proposal.  Since there seem to be implementations that, incorrectly, don't
remove the DH groups in the IKE_AUTH exchange (e.g. WatchGuard) we just strip
them from both proposals (it's a bit more lenient that way and we don't	have
to complicate the code to only clone and strip the local proposal, which
depends on a flag).